+++ chatlas/types/openai/_submit_responses.py
@@ -106,6 +106,11 @@ class SubmitInputArgs(TypedDict, total=False):
     model: Union[
         str,
         Literal[
+            "gpt-5.2",
+            "gpt-5.2-2025-12-11",
+            "gpt-5.2-chat-latest",
+            "gpt-5.2-pro",
+            "gpt-5.2-pro-2025-12-11",
             "gpt-5.1",
             "gpt-5.1-2025-11-13",
             "gpt-5.1-codex",